Replace pages with single controller file
