--- src/pages/compose.js
@@ -1,7 +1,0 @@
-'use strict'
-
-const composeView = require('./views/compose')
-
-module.exports = async function composePage () {
-  return composeView()
-}
