--- src/pages/hashtag.js
@@ -1,10 +1,0 @@
-'use strict'
-
-const post = require('./models/post')
-const listView = require('./views/list')
-
-module.exports = async function hashtag (channel) {
-  const messages = await post.fromHashtag(channel)
-
-  return listView({ messages })
-}
