--- src/pages/markdown.js
@@ -1,7 +1,0 @@
-'use strict'
-
-const markdownView = require('./views/markdown')
-
-module.exports = async function statusPage (text) {
-  return markdownView({ text })
-}
