--- src/pages/public-latest.js
@@ -1,9 +1,0 @@
-'use strict'
-
-const publicView = require('./views/public')
-const post = require('./models/post')
-
-module.exports = async function publicLatestPage () {
-  const messages = await post.latest()
-  return publicView({ messages })
-}
