--- src/pages/publish.js
@@ -1,15 +1,0 @@
-'use strict'
-
-const ssbMentions = require('ssb-mentions')
-const post = require('./models/post')
-
-module.exports = async function publishPage ({ text }) {
-  const mentions = ssbMentions(text).filter((mention) =>
-    mention != null
-  ) || undefined
-
-  return post.root({
-    text,
-    mentions
-  })
-}
